statfs: bugfix statfs02 fstatfs02 failed in ltp-testsuite.
In the test-case of statfs from ltp-testsuite, it pass -1 to 2th
argument. eg: fstatfs(fd , -1)

When uclibc-ng parse the buf32 to buf will cause illegal address
access, the kernel will signal the process with SIGSEV.

If we pass the -1 directly to the syscall of statfs/fstatfs, kenrel
use copy_to_user() to prevent the singal of SIGSEV and just return
EVINAL.

This is the ltp-testsuite expect.

Signed-off-by: Guo Ren <ren_guo@c-sky.com>